requirements.txt
python-telegram-bot==21.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
python-dotenv>=1.0.0
uuid6>=2024.1.12
